requirements.txt
python-telegram-bot==20.6
Flask==2.2.5
aiohttp==3.9.3
pandas==2.2.2
python-dotenv==1.0.1
psycopg2-binary
openpyxl
sqlalchemy
requests
msal
fastapi
uvicorn